restructure code to reduce cognitive load
