+++ signer-service/src/api/routes/v1/moonbeam.route.ts
@@ -2,7 +2,7 @@ import { Router } from 'express';
 import { executeXcmController } from '../../controllers/moonbeam.controller';
 import { validateExecuteXCM } from '../../middlewares/validators';
 
-const router: Router = Router({ mergeParams: true });
+const router: Router = Router();
 
 router.route('/execute-xcm').post(validateExecuteXCM, executeXcmController);
