+++ signer-service/src/api/routes/v1/pendulum.route.ts
@@ -1,7 +1,7 @@
 import { Router } from 'express';
 import { fundEphemeralAccountController } from '../../controllers/pendulum.controller';
 
-const router: Router = Router({ mergeParams: true });
+const router: Router = Router();
 
 router.post('/fundEphemeral', fundEphemeralAccountController);
