+++ signer-service/src/api/routes/v1/quote.route.ts
@@ -1,9 +1,9 @@
 import { Router } from 'express';
 import { getQuoteForProvider } from '../../controllers/quote.controller';
-import { validateQuoteInput } from '../../middlewares/validators';
+import { validateQuoteInput, QuoteQuery } from '../../middlewares/validators';
 
 const router: Router = Router({ mergeParams: true });
 
-router.route('/').get(validateQuoteInput, getQuoteForProvider);
+router.route('/').get<{}, unknown, {}, QuoteQuery>(validateQuoteInput, getQuoteForProvider);
 
 export default router;
